Refactor get_sequence_transform to utils and support multiple ignore indices
