+++ ignite/metrics/accuracy.py
@@ -264,38 +264,4 @@ class Accuracy(_BaseClassification):
             raise NotComputableError("Accuracy must have at least one example before it can be computed.")
         return self._num_correct.item() / self._num_examples
 
-    @staticmethod
-    def get_sequence_transform(
-        ignore_index: int | None = None,
-        output_transform: Callable = lambda x: x,
-    ) -> Callable:
-        """
-        Returns a callable to transform sequence model outputs for metric evaluation.
-        It flattens the sequences and filters out the padding (``ignore_index``).
-        """
-        def wrapper(output: Sequence[torch.Tensor]) -> tuple[torch.Tensor, torch.Tensor]:
-            y_pred, y = output_transform(output)
 
-            if y_pred.ndimension() == 3 and y.ndimension() == 2:
-                if y_pred.shape[:2] == y.shape:
-                    # y_pred is (N, L, C), y is (N, L)
-                    y_pred = y_pred.reshape(-1, y_pred.size(-1))
-                    y = y.reshape(-1)
-                elif y_pred.shape[0] == y.shape[0] and y_pred.shape[2] == y.shape[1]:
-                    # y_pred is (N, C, L), y is (N, L)
-                    y_pred = y_pred.transpose(1, 2).reshape(-1, y_pred.size(1))
-                    y = y.reshape(-1)
-            elif y_pred.ndimension() == 2 and y.ndimension() == 2:
-                # y_pred is (N, L), y is (N, L)
-                if y_pred.shape == y.shape:
-                    y_pred = y_pred.reshape(-1)
-                    y = y.reshape(-1)
-
-            if ignore_index is not None:
-                mask = y != ignore_index
-                y_pred = y_pred[mask]
-                y = y[mask]
-
-            return y_pred, y
-
-        return wrapper
