+++ tests/ignite/metrics/test_accuracy.py
@@ -500,44 +500,3 @@ def test_skip_unrolling():
     acc.iteration_completed(engine)
 
 
-def test_get_sequence_transform():
-    # test (N, L, C)
-    y_pred = torch.tensor(
-        [
-            [[0.1, 0.9], [0.8, 0.2], [0.3, 0.7], [0.5, 0.5]],
-            [[0.9, 0.1], [0.2, 0.8], [0.4, 0.6], [0.5, 0.5]],
-        ]
-    )  # shape: (2, 4, 2)
-    y = torch.tensor([[1, 0, 1, -1], [0, 1, 0, -1]])  # shape: (2, 4)
-
-    transform = Accuracy.get_sequence_transform(ignore_index=-1)
-    y_pred_t, y_t = transform((y_pred, y))
-
-    assert y_pred_t.shape == (6, 2)
-    assert y_t.shape == (6,)
-    assert y_t.tolist() == [1, 0, 1, 0, 1, 0]
-    assert y_pred_t[:, 1].tolist() == pytest.approx([0.9, 0.2, 0.7, 0.1, 0.8, 0.6])
-
-    # test (N, C, L)
-    y_pred_ncl = y_pred.transpose(1, 2).contiguous()  # (2, 2, 4)
-    y_pred_t2, y_t2 = transform((y_pred_ncl, y))
-    assert y_pred_t2.shape == (6, 2)
-    assert torch.all(y_pred_t2 == y_pred_t)
-    assert torch.all(y_t2 == y_t)
-
-    # test binary (N, L)
-    y_pred_bin = torch.tensor([[1, 0, 1, 1], [0, 1, 0, 0]])
-    y_bin = torch.tensor([[1, 0, 1, 2], [0, 1, 0, 2]])
-    transform_bin = Accuracy.get_sequence_transform(ignore_index=2)
-    y_pred_bin_t, y_bin_t = transform_bin((y_pred_bin, y_bin))
-    
-    assert y_pred_bin_t.shape == (6,)
-    assert y_bin_t.shape == (6,)
-    assert y_bin_t.tolist() == [1, 0, 1, 0, 1, 0]
-    assert y_pred_bin_t.tolist() == [1, 0, 1, 0, 1, 0]
-
-    # test without padding
-    transform_nopad = Accuracy.get_sequence_transform()
-    y_pred_nopad, y_nopad = transform_nopad((y_pred_bin, y_bin))
-    assert y_pred_nopad.shape == (8,)
-    assert y_nopad.shape == (8,)
